cleanup legacy code
