+++ plugin/src/main/kotlin/com/vanniktech/maven/publish/MavenPublishPlugin.kt
@@ -1,9 +1,13 @@
 package com.vanniktech.maven.publish
 
-import com.vanniktech.maven.publish.legacy.configurePlatform
-import com.vanniktech.maven.publish.legacy.setCoordinates
+import org.gradle.api.JavaVersion
 import org.gradle.api.Plugin
 import org.gradle.api.Project
+import org.gradle.api.plugins.JavaPluginExtension
+import org.gradle.api.publish.maven.MavenPublication
+import org.gradle.api.tasks.javadoc.Javadoc
+import org.gradle.external.javadoc.StandardJavadocDocletOptions
+import org.jetbrains.dokka.gradle.DokkaTask
 
 open class MavenPublishPlugin : Plugin<Project> {
 
@@ -12,7 +16,6 @@ open class MavenPublishPlugin : Plugin<Project> {
     val baseExtension = project.baseExtension
 
     project.setCoordinates()
-    project.configurePlatform()
 
     val sonatypeHost = project.findOptionalProperty("SONATYPE_HOST")
     if (sonatypeHost != null && sonatypeHost.isNotBlank()) {
@@ -24,5 +27,119 @@ open class MavenPublishPlugin : Plugin<Project> {
     }
 
     baseExtension.pomFromGradleProperties()
+
+    project.configurePlatform()
+  }
+}
+
+private fun Project.setCoordinates() {
+  group = project.findOptionalProperty("GROUP") ?: group
+  version = project.findOptionalProperty("VERSION_NAME") ?: version
+
+  // Artifact id defaults to project name which is not mutable.
+  // Some created publications use derived artifact ids (e.g. library, library-jvm, library-js) so it needs to be
+  // replaced instead of just set.
+  val artifactId = project.findOptionalProperty("POM_ARTIFACT_ID")
+  if (artifactId != null && artifactId != project.name) {
+    gradlePublishing.publications
+      .withType(MavenPublication::class.java)
+      .configureEach { publication ->
+        // skip the plugin marker artifact which has it's own artifact id based on the plugin id
+        if (publication.name.endsWith("PluginMarkerMaven")) {
+          return@configureEach
+        }
+
+        val projectName = name
+        val updatedArtifactId = publication.artifactId.replace(projectName, artifactId)
+        publication.artifactId = updatedArtifactId
+
+        // in Kotlin MPP projects some publications change our manually set artifactId again
+        afterEvaluate {
+          gradlePublishing.publications.withType(MavenPublication::class.java).named(publication.name).configure { publication ->
+            if (publication.artifactId != updatedArtifactId) {
+              publication.artifactId = publication.artifactId.replace(projectName, artifactId)
+            }
+          }
+        }
+      }
+  }
+}
+
+private fun Project.configurePlatform() {
+  plugins.withId("org.jetbrains.kotlin.multiplatform") {
+    baseExtension.configure(KotlinMultiplatform(defaultJavaDocOption() ?: JavadocJar.Empty()))
+  }
+
+  plugins.withId("com.android.library") {
+    // afterEvaluate is too late but we can't run this synchronously because we shouldn't call the APIs for
+    // multiplatform projects that use Android
+    androidComponents.finalizeDsl {
+      if (!plugins.hasPlugin("org.jetbrains.kotlin.multiplatform")) {
+        baseExtension.configure(AndroidMultiVariantLibrary())
+      }
+    }
+  }
+
+  afterEvaluate {
+    when {
+      plugins.hasPlugin("org.jetbrains.kotlin.multiplatform") -> {} // Handled above.
+      plugins.hasPlugin("com.android.library") -> {} // Handled above.
+      plugins.hasPlugin("java-gradle-plugin") ->
+        baseExtension.configure(GradlePlugin(defaultJavaDocOption() ?: javadoc()))
+      plugins.hasPlugin("org.jetbrains.kotlin.jvm") ->
+        baseExtension.configure(KotlinJvm(defaultJavaDocOption() ?: javadoc()))
+      plugins.hasPlugin("org.jetbrains.kotlin.js") ->
+        baseExtension.configure(KotlinJs(defaultJavaDocOption() ?: JavadocJar.Empty()))
+      plugins.hasPlugin("java-library") ->
+        baseExtension.configure(JavaLibrary(defaultJavaDocOption() ?: javadoc()))
+      plugins.hasPlugin("java") ->
+        baseExtension.configure(JavaLibrary(defaultJavaDocOption() ?: javadoc()))
+      else -> logger.warn("No compatible plugin found in project $name for publishing")
+    }
+  }
+}
+
+private fun Project.defaultJavaDocOption(): JavadocJar? {
+  return if (plugins.hasPlugin("org.jetbrains.dokka") || plugins.hasPlugin("org.jetbrains.dokka-android")) {
+    JavadocJar.Dokka(findDokkaTask())
+  } else {
+    null
+  }
+}
+
+private fun Project.javadoc(): JavadocJar {
+  tasks.withType(Javadoc::class.java).configureEach {
+    val options = it.options as StandardJavadocDocletOptions
+    val javaVersion = javaVersion()
+    if (javaVersion.isJava9Compatible) {
+      options.addBooleanOption("html5", true)
+    }
+    if (javaVersion.isJava8Compatible) {
+      options.addStringOption("Xdoclint:none", "-quiet")
+    }
+  }
+  return JavadocJar.Javadoc()
+}
+
+private fun Project.javaVersion(): JavaVersion {
+  try {
+    val extension = project.extensions.findByType(JavaPluginExtension::class.java)
+    if (extension != null) {
+      val toolchain = extension.toolchain
+      val version = toolchain.languageVersion.forUseAtConfigurationTime().get().asInt()
+      return JavaVersion.toVersion(version)
+    }
+  } catch (t: Throwable) {
+    // ignore failures and fallback to java version in which Gradle is running
+  }
+  return JavaVersion.current()
+}
+
+private fun Project.findDokkaTask(): String {
+  val tasks = project.tasks.withType(DokkaTask::class.java)
+  return if (tasks.size == 1) {
+    tasks.first().name
+  } else {
+    tasks.findByName("dokkaHtml")?.name ?: "dokka"
   }
 }
